Temp Kludge get get VICE's tcp/ip RS232 ports working again
Instead of properly caching the virtual tcp/ip RS232 connection, VICE attempts to re-open the same connection multiple times as the internal code opens the RS232 port.  tcpser is properly designed to reject connections when already connected to a VICE instance, which causes issues (VICE wants to send data on the new connection, while tcpser is expecting it on the old one.  This kludge drops the old connection and switches to the new one, which is a major security risk (someone can then hijack the tcpser ip232 connection much more easily), but this temp solution may address the immediate symptom until VICE cna be patched.